lambda_fns/passports/requirements.txt
boto3==1.36.6
requests==2.32.3
pyjwt[crypto]==2.10.1